requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
filelock==3.9.0
fsspec==2023.12.2
huggingface-hub==0.20.2
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
packaging==23.2
Pillow==9.3.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.28.1
safetensors==0.4.1
sympy==1.12
tokenizers==0.15.0
torch==2.1.2+cpu
torchaudio==2.1.2+cpu
torchvision==0.16.2+cpu
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.4.0
urllib3==1.26.13
